Composition
* Add components package
* Add HeaderPage for navigation
* Add ProductThumbnail for adding to cart
