+++ master-selenium-framework/src/test/java/dev/ericmarcelo/selenium/pom/tests/AddToCartTest.java
@@ -7,13 +7,8 @@ import dev.ericmarcelo.selenium.pom.pages.CartPage;
 import dev.ericmarcelo.selenium.pom.pages.HomePage;
 import dev.ericmarcelo.selenium.pom.pages.ProductPage;
 import dev.ericmarcelo.selenium.pom.pages.StorePage;
-import dev.ericmarcelo.selenium.pom.utils.ProductsJacksonUtils;
 import org.testng.Assert;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,6 +18,7 @@ public class AddToCartTest extends BaseTest {
     public void addToCartFromStorePage() {
         CartPage cartPage = new StorePage(getWebDriver())
                 .load()
+                .getProductThumbnail()
                 .clickAddToCartButton("Blue Shoes")
                 .clickViewCart();
         Assert.assertEquals(cartPage.getProductName(), "Blue Shoes");
@@ -46,7 +42,9 @@ public class AddToCartTest extends BaseTest {
         List<Product> featuredProducts = new ProductsDataProvider().getFeaturedProducts();
         HomePage homePage = new HomePage(getWebDriver());
         Optional<Product> optionalProduct = featuredProducts.parallelStream().findAny();
-        CartPage cartPage = homePage.load()
+        CartPage cartPage = homePage
+                .load()
+                .getProductThumbnail()
                 .clickAddToCartButton(optionalProduct.get().getName())
                 .clickViewCart();
         Assert.assertEquals(cartPage.getProductName(), optionalProduct.get().getName());
