+++ master-selenium-framework/src/test/java/dev/ericmarcelo/selenium/pom/tests/MyFirstTestCase.java
@@ -31,12 +31,12 @@ public class MyFirstTestCase extends BaseTest {
         BillingJacksonUtils.deserializeBillingAddressJson(is, billingAddress);
 
         StorePage storePage = new HomePage(getWebDriver())
-                .load().navigateToStoreUsingMenu()
+                .load().getHeaderPage().navigateToStoreUsingMenu()
                 .search("Blue");
         Assert.assertEquals(storePage.getTitle(), "Search results: “Blue”");
 
-        storePage.clickAddToCartButton("Blue Shoes");
-        CartPage cartPage = storePage.clickViewCart();
+        storePage.getProductThumbnail().clickAddToCartButton("Blue Shoes");
+        CartPage cartPage = storePage.getProductThumbnail().clickViewCart();
         Assert.assertEquals(cartPage.getProductName(), "Blue Shoes");
 
         CheckoutPage checkoutPage = cartPage.checkout();
@@ -52,12 +52,12 @@ public class MyFirstTestCase extends BaseTest {
     public void loginAndCheckoutUsingDirectBankTransfer() {
 
         StorePage storePage = new HomePage(getWebDriver())
-                .load().navigateToStoreUsingMenu()
+                .load().getHeaderPage().navigateToStoreUsingMenu()
                 .search("Blue");
         Assert.assertEquals(storePage.getTitle(), "Search results: “Blue”");
 
-        storePage.clickAddToCartButton("Blue Shoes");
-        CartPage cartPage = storePage.clickViewCart();
+        storePage.getProductThumbnail().clickAddToCartButton("Blue Shoes");
+        CartPage cartPage = storePage.getProductThumbnail().clickViewCart();
         Assert.assertEquals(cartPage.getProductName(), "Blue Shoes");
 
         CheckoutPage checkoutPage = cartPage.checkout();
