+++ master-selenium-framework/src/test/resources/drivers/resolution.properties
@@ -1,8 +1,8 @@
 #WebDriverManager Resolution Cache
-#Mon Mar 27 12:17:00 PDT 2023
+#Mon Mar 27 13:59:55 PDT 2023
 firefox110-ttl=14\:34\:51 23/03/2023 PDT
 firefox111-ttl=10\:08\:50 28/03/2023 PDT
 firefox110=0.32.2
 firefox=111
 firefox111=0.32.2
-firefox-ttl=13\:17\:00 27/03/2023 PDT
+firefox-ttl=14\:59\:55 27/03/2023 PDT
